% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenos.R
\name{allGenos}
\alias{allGenos}
\title{Find all possible genotypes for one individual}
\usage{
allGenos(alleles)
}
\arguments{
\item{alleles}{Vector of allele names}
}
\value{
Returns matrix of all possible genotypes, one row per genotype.
}
\description{
Generates all genotypes
}
