% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenosPair.R
\name{allGenosPair}
\alias{allGenosPair}
\title{Find all possible genotypes for one individual}
\usage{
allGenosPair(alleles)
}
\arguments{
\item{alleles}{Vector of allele names}
}
\value{
Returns matrix of all possible genotypes, one row per genotype.
}
\description{
Generates all genotypes
}
\examples{
allGenos(1:2)
}
