% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPairsTable.R
\name{allPairsTable}
\alias{allPairsTable}
\title{Blind search table}
\usage{
allPairsTable(x, ids = NULL, markers = NULL)
}
\arguments{
\item{x}{A `ped` object or a list of such.}

\item{ids}{Either a vector with ID labels, or a data frame/matrix with two
columns, where each row contains the ID labels of two individuals. The entries
are coerced to characters, and must match uniquely against the ID labels of
`x`. If `ids` is a vector, it is converted to a matrix containing all pairs.
By default, all individuals of `x` are included.}

\item{markers}{A numeric indicating which marker(s) to include. If NULL
(default), all markers are used.}
}
\value{
A list with two elements, one data frame  for LRs, one for posteriors
with flat prior. The  data frame with 6 columns: `ID1`, `ID2`, `N` (the number of markers
with no missing alleles), `andLRs or posteriors.
}
\description{
Calculate table ofLRs and posteriors for PO, SI, SI, FC and UN for all pairs
}
\details{
A uniform priorisused so far
}
\examples{

### Example 1
library(forrel)
x = nuclearPed(2)
# Simulate 200 equifrequent SNPs
x = markerSim(x, N = 200, alleles = 1:2, verbose = FALSE)

# Parent offspring against unrelated
tab1 = allPairsTable(x)

### Example 2 X-markers
library(pedtools, quietly = T)
x = nuclearPed(2)
x = setMarkers(x, locusAttributes = dbX12)
chrom(x, 1:length(dbX12)) = "X"
x = profileSim(x, N = 10)
tab1 = allPairsTable(x[[1]])

}
\references{
To appear
}
\author{
Thore Egeland
}
