% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bi_matrices.R
\name{Bi_matrices}
\alias{Bi_matrices}
\title{Computes Bi, the 9 matrices in the variance formulain appendix of Brustad et al.}
\usage{
Bi_matrices(p)
}
\arguments{
\item{p}{Vector of allele frequencies}
}
\value{
List of matrices Bi
}
\description{
Computes Bi, the 9 matrices in the variance formulain appendix of Brustad et al.
}
\examples{
\dontrun{
Bi_matrices(c(0.2, 0.8))
}
}
