% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPhi.R
\name{bootPhi}
\alias{bootPhi}
\title{Properties of bootstrap confidence intervals for the kinship coefficient}
\usage{
bootPhi(
  ped,
  ids = NULL,
  N = 2,
  B = 2,
  CItype = "bca",
  conf.level = 0.95,
  plot = F,
  seed = NULL,
  verbose = F
)
}
\arguments{
\item{ped}{ped object with allele frequencies.}

\item{ids}{Id of pair.}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{CItype}{Logical. See [kinBoot::interval()]}

\item{conf.level}{Double}

\item{plot}{Logical}

\item{seed}{Integer.}

\item{verbose}{Logical.}
}
\value{
A list with the following four elements:

  * `phi`: The kinship coefficient of the pedigree.

  * `averaged` : A data frame with two lines, one for `parametric`
  and one for `nonParametric`. The values are averaged over the `N` simulations.
  The columns `realised`, `boot`, `skew`, `lower`,  `upper`, and  `cover` are explained
  in `Details.`

  * `simParametric`. The entries are as for `averaged`
  but for each parametric simulation.

  * `simNonparametric` The entries are as for `averaged`
  but for each nonparametric simulation.
}
\description{
The purpose is to compare parametric and nonparametric confidence
intervals in kinship applications, currently only for the kinship coefficient.
}
\details{
Marker data are simulated `N` times giving `N` estimates of (kappa0, kappa1, kappa2).
For each simulation, the `realised` phi is found. Parametric and nonparametric bootstrapping
is done giving `boot` (the averaged kinship coefficient from `B` bootstrap simulations)
and `skew` = `realised`- `boot`. There are various
ways to calculate the confidence interval. The default is `bca` as implemented in `coxed::bca`. The
variable `cover` is 1 if the confidence interval contains `phi` and 0 otherwise.
}
\examples{
# The example considers the kinship coefficients between
# brothers named `B1` and `B2` using the 35 markers in
# forrel::NorwegianFrequencies.

library(forrel)
library(ribd)
library(coxed) # for bca confidence intervals

ids = c("B1", "B2")
ped = nuclearPed(2, children = ids)
ped = setMarkers(ped, locusAttributes = NorwegianFrequencies)
N = 10 # no of confidence intervals. Increase
B = 100 # no of  bootstraps. Increase
res1 = bootPhi(ped, ids, N, B, seed = 17)

# Basic output
res1[1:2]

# Compare parametric and nonparametric estimates
y1 = res1$simParametric$boot
y2 = res1$simNonparametric$boot
boxplot(y1, y2, names = c("parametric", "nonparametric"),
        main = "Bootstrap estimates of kinship coefficient",
        sub = "Red stapled line: theoretical value")
abline(h = res1$phi, col = 'red', lty = 2)
}
