% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{case}
\alias{case}
\title{Case determination for a pair of individuals}
\usage{
case(a, b, cc, d)
}
\arguments{
\item{a}{Vector of positive integers (allele 1 of individual 1)}

\item{b}{Vector of positive integers (allele 2 of individual 1)}

\item{cc}{Vector of positive integers (allele 1 of individual 2)}

\item{d}{Vector of positive integers (allele 2 of individual 2)}
}
\value{
`case`
}
\description{
This is a reduced version of the function `likJ`. It only returns
the 'casew' variable oflikJ
}
\examples{
# Below I go through some cases trying to figure out
# what the LD probabilities should be

p = c(1,2,3,4,5)/sum(1:5)
H = p\%o\%p # Haplotypeprobs for LE

### Case  1,1
a  = c(1, 2)
b  = c(1, 2)
cc = c(1, 2)
d  = c(1, 2)
(Case = case(a,b,cc,d))
pa = p[a]
pb = p[b]
pc = p[cc]
pd = p[d]
Delta = c(1, 0, 0, 0, 0, 0, 0, 0, 0)
lik.LE = prod(likJ(a, b, cc, d, pa, pb, pc, pd, Delta))
lik.LD = H[a[1], a[2]]
lik.LD = geno1(cc[1], d[1], cc[2], d[2], H)
abs(lik.LE -lik.LD) < 1e-15

### Case 1,2
a  = c(1, 1)
b  = c(1, 1)
cc = c(1, 2)
d  = c(1, 2)
(Case = case(a,b,cc,d))
pa = p[a]
pb = p[b]
pc = p[cc]
pd = p[d]
Delta1 = c(1, 0, 0, 0, 0, 0, 0, 0, 0)
Delta2 = c(0, 1, 0, 0, 0, 0, 0, 0, 0)
lik.LE = likJ(a[1], b[1], cc[1], d[1], pa[1], pb[1], pc[1], pd[1], Delta1)*
         likJ(a[2], b[2], cc[2], d[2], pa[2], pb[2], pc[2], pd[2], Delta2)
# P(G1,G2|Jst)= P(G1|Jst, G2)  * P(G2|Jst) = p[a[[1]]]*H[cc[1],cc[2]]
lik.LD = p[a[[2]]]*H[cc[1],cc[2]]
lik.LD = p[a[[2]]]*geno1(cc[1], d[1], cc[2], d[2], H)
abs(lik.LE -lik.LD) < 1e-15

### Case 2,1
a  = c(1, 1)
b  = c(1, 1)
cc = c(2, 1)
d  = c(2, 1)
(Case = case(a,b,cc,d))
pa = p[a]
pb = p[b]
pc = p[cc]
pd = p[d]
Delta1 = c(0, 1, 0, 0, 0, 0, 0, 0, 0)
Delta2 = c(1, 0, 0, 0, 0, 0, 0, 0, 0)
lik.LE = likJ(a[1], b[1], cc[1], d[1], pa[1], pb[1], pc[1], pd[1], Delta1)*
         likJ(a[2], b[2], cc[2], d[2], pa[2], pb[2], pc[2], pd[2], Delta2)
# P(G1,G2|Jst)= P(G1|Jst, G2)  * P(G2|Jst) = p[a[[1]]]*H[cc[1],cc[2]]
lik.LD = p[a[[1]]]*H[cc[1],cc[2]]
lik.LD = p[a[[1]]]*geno1(cc[1], d[1], cc[2], d[2], H)
abs(lik.LE -lik.LD) < 1e-15

}
