% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCI.R
\name{compareCI}
\alias{compareCI}
\title{Compare bootstrap CI-s (nonparametric, parametric) and assymptotic}
\usage{
compareCI(theta, x, ids, n = 2, N = 2, B = 2, seed = NULL)
}
\arguments{
\item{theta}{Double kappa0.}

\item{x}{ped obects}

\item{ids}{Id of pair}

\item{n}{Integer. No of markers}

\item{N}{Integer. No of simulations}

\item{B}{Integer. No of bootstraps}
}
\value{
Returns table.
}
\description{
The
}
\examples{
library(forrel)
library(pedprobr)
x = halfSibPed(1)
ids = c(4,5)
n = 100
N = 100
B = 10
theta = 0.5
seed = 17
compareCI(theta, x, ids, n, N, B, seed = seed)
}
