% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragePhi.R
\name{coveragePhi}
\alias{coveragePhi}
\title{Coverage for bootstrap confidence intervals}
\usage{
coveragePhi(kappa, ped, ids = NULL, N = 2, B = 2, seed = NULL, level = 0.95)
}
\arguments{
\item{kappa}{Probability vector of length 3.}

\item{ped}{ped object with allele frequencies.}

\item{ids}{Id of pair.}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{seed}{Integer.}

\item{level}{Double}
}
\value{
Returns a l
}
\description{
Currently we only consider parametric bootstrap and the kinship coefficient.
}
\details{
The equations
}
\examples{
library(forrel)
library(pedprobr)
library(ribd)
library(moments)

# Example 1
n = 1000 # no of markers
p = rep(0.5, n)
freq = list()
for (i in 1:n)
  freq[[i]] =  list(afreq = c("1" = p[i], "2" = 1- p[i]))
ped = quadHalfFirstCousins()
ped = setMarkers(ped, locusAttributes = freq)
N = 400 # no of confidence intervals
B = 400 # no of  bootstraps
seed = 2
kappa = ribd::kappaIBD(ped, ids = leaves(ped))
res1 = coveragePhi(kappa = kappa, ped = ped, ids, N = N, B = B, seed = seed, level = 0.95)
round(res1[c(1:2,N+1),], 4)
}
