% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePhi.R
\name{examplePhi}
\alias{examplePhi}
\title{Bootstrap confidence intervals}
\usage{
examplePhi(
  peds,
  idlist,
  N = 2,
  B = 10,
  CItype = "bca",
  conf.level = 0.95,
  seed = NULL,
  verbose = TRUE,
  db = NULL
)
}
\arguments{
\item{peds}{list of ped objects with allele frequencies.}

\item{idlist}{list of ids of of pair.}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{CItype}{Logical}

\item{conf.level}{Double}

\item{seed}{Integer}

\item{verbose}{Logical}

\item{db}{database}
}
\value{
Returns a data frame summarising the simulation.
}
\description{
We consider parametric and nonparametric bootstrap and the kinship coefficient for some pedigrees.
}
\details{
See [kinBoot::bootPhi()]
}
\examples{
library(forrel)
library(ribd)
library(coxed) # for bca confidence intervals

peds = list(quadHalfFirstCousins(), doubleFirstCousins(), nuclearPed(2),
            halfSibPed(), cousinPed(1))
names(peds) = c("QHFC",  "DFC", "S", "H", "FC")
phi = unlist(lapply(peds, function(x) kinship(x, leaves(x))))
idlist = lapply(peds, leaves)
N = 1; B = 1000; seed = 1729
# Example 1 Many snps
n = 1000 # no of markers
p = rep(0.5, n)
freq = list()
for (i in 1:n)
freq[[i]] =  list(afreq = c("1" = p[i], "2" = 1- p[i]))
db = freq
res1 = examplePhi(peds, idlist, N = N, B = B, seed = seed, db = db)

# Example 2 NorwegianFrequency
db = NorwegianFrequencies
res2 = examplePhi(peds, idlist, N = N, B = B, seed = seed, db = db)

# Example 3 Few markers, 8 CODIS markers
codis8 = c("CSF1PO", "D3S1358", "D5S818", "D7S820", "D8S1179", "D13S317", "D16S539", "D18S51")
db = NorwegianFrequencies[codis8]
res3 = examplePhi(peds, idlist, N = N, B = B, seed = seed, db = db)
res = rbind(res1, res2, res3)
res
}
