% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno1.R
\name{geno1}
\alias{geno1}
\title{Calculates P(G|H)}
\usage{
geno1(a1, a2, b1, b2, H)
}
\arguments{
\item{a1}{First allele, first marker}

\item{a2}{Second allele, first marker}

\item{b1}{First allele, second marker}

\item{b2}{Second allele, second marker}
}
\value{
P(G|H)
}
\description{
The marker data for one individual at two markers is
G = (a1/a2, b1/b2). H is a n by n matrix of haplotype probabilities
}
\examples{
# Below I go through some cases trying to figure out
# what the LD probabilities should be

p = q = c(0.1, 0.2, 0.3, 0.4)
# Checks when LE
H = p\%o\%q # Haplotype probs for LE

abs(geno1(1, 1, 1, 1, H) - p[1]^2*q[1]^2) < 1e-15
abs(geno1(1, 2, 3, 3, H) - 2*p[1]*p[2]*q[3]^2) < 1e-15
abs(geno1(1, 2, 3, 4, H) - 2*p[1]*p[2]*2*q[3]*q[4]) < 1e-15
abs(geno1(2, 2, 4, 4, H) - p[2]^2*q[4]^2) < 1e-15
}
