% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{interval}
\alias{interval}
\title{Bootstrap confidence intervals}
\usage{
interval(x, method = "bca", conf.level = 0.95)
}
\arguments{
\item{x}{Double vector. Bootstrap sample.}

\item{method}{Character. See details.}

\item{conf.level}{Double in(0,1). Confidence level}
}
\value{
A vector of length 2
in which the first element is the lower bound and the second element is the upper bound.
}
\description{
Some alternative confidence intervals are implemented.
}
\details{
The implemented methods are `perc`, the standard using the `quantile` function R
to find the percentile; `bca`, see `coxed::bca``, and 'student' based on
the normal approximation.
}
\examples{
library(coxed)
x = rnorm(100000)
interval(x, method = "perc")
interval(x, method = "bca")
interval(x, method = "student")
interval(x, method = "Not implemented")

# Example Efron & Tibshirani p 154
x = c(52, 104, 146, 10, 51, 30, 40, 27, 46)
interval(x, method = "bca", conf.level = 0.9)
}
