% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likJ.R
\name{likJ}
\alias{likJ}
\title{Likelihood for two Possibly Inbred Individuals as a function of
the Condensed 'Jacquard' Coefficients}
\usage{
likJ(a, b, cc, d, pa, pb, pc, pd, Delta)
}
\arguments{
\item{a}{Vector of positive integers (allele 1 of individual 1)}

\item{b}{Vector of positive integers (allele 2 of individual 1)}

\item{cc}{Vector of positive integers (allele 1 of individual 2)}

\item{d}{Vector of positive integers (allele 2 of individual 2)}

\item{pa}{Double vector of allele frequencies}

\item{pb}{Double vector of allele frequencies}

\item{pc}{Double vector of allele frequencies}

\item{pd}{Double vector of allele frequencies}

\item{Delta}{Double vector of length 9 summing to unity}
}
\description{
Likelihood for two Possibly Inbred Individuals as a function of
the Condensed 'Jacquard' Coefficients
}
\examples{
p = c("1" = 0.1, "2" = 0.9)
a  = c(rep(1,6), rep(2,3))
b  = c(rep(1,3), rep(2,6))
cc = rep(c(1,1,2), 3)
d  = rep(c(1,2,2), 3)
pa = p[a]
pb = p[b]
pc = p[cc]
pd = p[d]
Delta = c(0, 0, 0, 0, 0, 0, 1,0, 0)
l2 = likJ(a,b,cc,d, pa,pb,pc,pd, Delta = Delta)
sum(l2) == 1
}
