% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likJStates.R
\name{likJStates}
\alias{likJStates}
\title{Likelihood for for all 9 Jacquard states}
\usage{
likJStates(a, b, cc, d, pa, pb, pc, pd)
}
\arguments{
\item{a}{Vector of positive integers (allele 1 of individual 1)}

\item{b}{Vector of positive integers (allele 2 of individual 1)}

\item{cc}{Vector of positive integers (allele 1 of individual 2)}

\item{d}{Vector of positive integers (allele 2 of individual 2)}

\item{pa}{Double vector of allele frequencies}

\item{pb}{Double vector of allele frequencies}

\item{pc}{Double vector of allele frequencies}

\item{pd}{Double vector of allele frequencies}
}
\value{
A matrix with 9 rows and columns equal to the number of markers.
Element (i,j) is the likelihood conditioned on Jacquard state i for marker j.
}
\description{
Likelihood for for all 9 Jacquard states
}
\examples{
p = c(0.1, 0.2, 0.3, 0.4)
a  = c(1,1)
b  = c(2,2)
cc = c(3,1)
d  = c(4,2)
pa = p[a]
pb = p[b]
pc = p[cc]
pd = p[d]
Delta = c(0, 0, 0, 0, 0, 0, 1,0, 0) #unrelated
l2 = likJStates(a,b,cc,d, pa,pb,pc,pd)
}
