% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRH.R
\name{LRH}
\alias{LRH}
\title{Calculate the 9x9 matrix used to find expected value}
\usage{
LRH(p)
}
\arguments{
\item{p}{Double vector, allele frequencies}
}
\description{
See inbred.pdf note
}
\examples{
LRH(c(0.2, 0.8))

}
