% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreePhi.R
\name{pedigreePhi}
\alias{pedigreePhi}
\title{Coverage for bootstrap confidence intervals}
\usage{
pedigreePhi(
  kappa,
  ped,
  ids = NULL,
  N = 2,
  B = 2,
  seed = NULL,
  level = 0.95,
  plot = F
)
}
\arguments{
\item{kappa}{Probability vector of length 3.}

\item{ped}{ped object with allele frequencies.}

\item{ids}{Id of pair.}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{seed}{Integer.}

\item{level}{Double}

\item{plot}{Logical}
}
\value{
Returns a l

Returns a list with two elements, a dataframe and the result
  of last simulation called `lastBoot`(see documentation of `ibdBootstrap`).
  The columns of the data frame are

  * `phi.hat` The estimate for each bootstrap

  * `skew` See details

  * `dist` See documentation of `ibdBootstrap`

  * `lower` Lower bound for interval


  * `upper` Upper bound for interval

  * `coverage`
}
\description{
Currently we only consider parametric bootstrap and the kinship coefficient.
}
\details{
See note coverage.pdf
}
\examples{
library(forrel)
library(pedprobr)
library(ribd)
library(moments)

# Example Estimate coverage and more
n = 100 # no of markers
p = rep(0.5, n)
freq = list()
for (i in 1:n)
  freq[[i]] =  list(afreq = c("1" = p[i], "2" = 1- p[i]))
ped = quadHalfFirstCousins()
ped = setMarkers(ped, locusAttributes = freq)
N = 2 # no of confidence intervals
B = 100 # no of  bootstraps
seed = 2
kappa = ribd::kappaIBD(ped, ids = leaves(ped))
res1 = pedigreePhi(kappa = kappa, ped = ped, ids,
                   N = N, B = B, seed = seed, level = 0.95, plot = T)
round(res1$tab[c(1:2,N+1),], 4)
boot1 = res1$lastBoot
phi.hat = 0.25*boot1$k1 + 0.5*boot1$k2
plot(density(phi.hat), main = "", xlab = "Kinship coefficient")
qqnorm(phi.hat)
}
