% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankProfiles.R
\name{rankProfiles}
\alias{rankProfiles}
\title{The possible profiles for a member of pedigree are probability ranked}
\usage{
rankProfiles(x, id, markers = NULL, maxPerMarker = Inf, verbose = FALSE)
}
\arguments{
\item{id}{Name of individual to be predicted.}

\item{maxPerMarker}{No of top candidates per marker, default all.}

\item{Names}{or indices of the markers to be included. By default, all markers.}

\item{A}{logical, by default TRUE.}
}
\value{
A list with three components. The first, a data frame with profiles
ranked according to the likelihood. The posterior equals the likely for a prior which
gives equal probability to all possible profiles. Then follows the most likely profile. Finally,
the second to most likely genotypes are given.
}
\description{
The possible profiles for a member of pedigree are probability ranked
}
\details{
Markers are assumed independent. For each marker, the possible genotypes
for `id` are evaluated and ranked according to how probable they are. It may be wise
to try first with `maxPerMarker = 1` to limit computation time, particularly if mutations are modelled.
}
\examples{
library(pedprobr)
x = nuclearPed(2, father = "FA")
m1 = marker(x, "3" = 1, "4" = 1:2, alleles = 1:2, name = "m1")
m2 = marker(x, "3" = 1, "4" = 2, alleles = 1:2, name = "m2")
m3 = marker(x, "3" = 2, "4" = 2, alleles = 1:2, name = "m3")
x = setMarkers(x, list(m1, m2, m3))
id = "FA"
markers = NULL
maxPerMarker = 2
rankProfiles(x, id, markers, maxPerMarker)
rankProfiles(x, id, markers, maxPerMarker = 1)
library(pedmut)
mutmod(x, 1:3) = list(model = "equal", rate = 0.002)
rankProfiles(x, id, markers, maxPerMarker = Inf)
}
\author{
Magnus Dehli Vigeland
}
