% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDatabase.R
\name{readDatabase}
\alias{readDatabase}
\title{Reads database of allele frequencies}
\usage{
readDatabase(filename)
}
\arguments{
\item{character}{filename}
}
\description{
Reads database of allele frequencies
}
\details{
An tab separated asci file is read and converted to
`ped` format. The first line gives the name of of the marker,
the next lines the allele name and frequency. The markers are
separated by blank lines.
}
\examples{
\dontrun{
library(pedtools)
db = readDatabase("markerDB.txt")
# Attach to a pedigree
x = setMarkers(singleton(1), locusAttributes = db)
# Specify X chromosome
chrom(x, 1:length(db)) = "X"
x
}
}
