% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realisedPhi.R
\name{realisedPhi}
\alias{realisedPhi}
\title{Bootstrap intervals for realised phi}
\usage{
realisedPhi(
  ped,
  ids = NULL,
  N = 2,
  B = 2,
  seed = NULL,
  level = 0.95,
  method = "parametric",
  bca = FALSE
)
}
\arguments{
\item{ped}{ped object with allele frequencies.}

\item{ids}{Id of pair.}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{seed}{Integer.}

\item{level}{Double}

\item{method}{Character 'parametric' or 'nonparametric'}

\item{bca}{Logical}
}
\value{
Returns a l
}
\description{
Currently we only consider parametric bootstrap and the kinship coefficient.
}
\details{
The equations
}
\examples{
library(forrel)
library(pedprobr)
library(ribd)
library(moments)
library(coxed)

# Example 1a Estimate coverage and more
n = 100 # no of markers
p = rep(0.5, n)
freq = list()
for (i in 1:n)
  freq[[i]] =  list(afreq = c("1" = p[i], "2" = 1- p[i]))
ped = quadHalfFirstCousins()
ids = leaves(ped)
ped = setMarkers(ped, locusAttributes = NorwegianFrequencies)
N = 100
B = 400
seed = 17
bca = FALSE
date()
resParametric = realisedPhi(ped = ped, ids = ids, N = N, B = B,
                            method = "parametric", seed = seed, level = 0.95, bca = bca)
resNonparametric = realisedPhi(ped = ped, ids = ids, N = N, B = B,
                            method = "nonparametric", seed = seed, level = 0.95, bca = bca)
res = rbind(resParametric[N+1,], resNonparametric[N+1,] )
rownames(res) = c("parametric", "nonparametric")
round(res,4)
date()

}
