% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPhi.R
\name{bootPhi}
\alias{bootPhi}
\title{Coverage for bootstrap confidence intervals}
\usage{
bootPhi(
  ped,
  ids = NULL,
  N = 2,
  B = 2,
  CItype = "bca",
  conf.level = 0.95,
  plot = F,
  seed = NULL,
  verbose = F
)
}
\arguments{
\item{ped}{ped object with allele frequencies.}

\item{ids}{Id of pair.}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{CItype}{Logical}

\item{conf.level}{Double}

\item{plot}{Logical}

\item{seed}{Integer.}

\item{verbose}{Logical.}
}
\value{
Returns a l

Returns a list with two elements, a dataframe and the result
  of last simulation called `lastBoot`(see documentation of `ibdBootstrap`).
  The columns of the data frame are

  * `phi.hat` The estimate for each bootstrap

  * `skew` See details

  * `dist` See documentation of `ibdBootstrap`

  * `lower` Lower bound for interval


  * `upper` Upper bound for interval

  * `coverage`
}
\description{
Currently we only consider parametric bootstrap and the kinship coefficient.
}
\details{
See note coverage.pdf
}
\examples{
library(forrel)
library(pedprobr)
library(ribd)
library(coxed)

# Example Estimate coverage and more
n = 100 # no of markers
p = rep(0.5, n)
freq = list()
for (i in 1:n)
  freq[[i]] =  list(afreq = c("1" = p[i], "2" = 1- p[i]))
ped = halfSibPed()
ped = setMarkers(ped, locusAttributes = NorwegianFrequencies)
ids = leaves(ped)
N = 100 # no of confidence intervals
B = 400 # no of  bootstraps
CItype = "bca"
conf.level = 0.8
plot = FALSE
seed = 2
res1 = bootPhi(ped = ped, ids = ids, N = N, B = B,
               CItype = CItype, conf.level = conf.level, plot = plot, seed = seed)

boot1 = res1$lastBoot
phi.hat = 0.25*boot1$k1 + 0.5*boot1$k2
plot(density(phi.hat), main = "", xlab = "Kinship coefficient")
qqnorm(phi.hat)
}
