% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCI.R
\name{compareCI}
\alias{compareCI}
\title{Compare kinskip confidence intervals}
\usage{
compareCI(
  theta,
  x,
  ids,
  n = NULL,
  N = 2,
  B = 2,
  seed = NULL,
  asymptotic = TRUE
)
}
\arguments{
\item{theta}{Double kappa0.}

\item{x}{ped object with allele frequencies.}

\item{ids}{Id of pair.}

\item{n}{Integer. No of markers, only used in the asymptoic case}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{seed}{Integer.}
}
\value{
Returns a list with length 3 or 4 (if asymptotic is included).
The last element of the list contains averaged CI, point estimate, coverage, and dist
(as defined in `ibdBootstrap`),
the former elements contain the same information for each CI.
}
\description{
The purpose is to compare confidence intervals (CI-s) based on parametric bootstrap and
nonparametric bootstrap. For a simple case, iid SNPs
with frequency 0.5 and kappa2 = 0, the asymptotic counterparts are provided.
}
\details{
The equations for the simple asymptotic case are documented separately.
CI-s for the bootstrap are calculated using the percentile method.
Only estimates in the asymptotic case are constrained by kappa2 = 0.
This makes comparison between the asymptotic and the bootstrap estimates a bit unfair.
}
\examples{
library(forrel)
library(pedprobr)

# Example 1. Assumptions for simple asymptotics met.
x = halfSibPed(1)
n = 100 # no of markers
N = 2 # no of simulations
B = 100 # no of  bootstraps

## Frequencies for bootstrap
freq = list()
for (i in 1:n)
 freq[[i]] =  list(alleles = 1:2, afreq = c(0.5, 0.5))
x = setMarkers(x, locusAttributes = freq)

## Parent offspring
ids = c(1, 4); theta = 0; seed = 17
foo1 = compareCI(theta, x, ids, n, N, B)

## half sibs
ids = c(4,5); theta = 0.5; seed = 17;
foo2 = compareCI(theta, x, ids, n, N, B,)

## Unrelated
ids = c(1,3); theta = 1
foo3 = compareCI(theta, x, ids, n, N, B)

cbind(theta = rep(c(0,0.5,1), each = 3),
      rbind(foo1$average, foo2$average, foo3$average))

# Example 2. Strong deviation from iid

n1 = 19 # no of id markers
n2 = 1
MAF = 0.001 #minor allele frequency
n = n1 + n2 # no of markers
N = 2 # no of simulations
B = 100 # no of  bootstraps

## Frequencies for bootstrap
freq = list()
for (i in 1:n1)
 freq[[i]] =  list(alleles = 1:2, afreq = c(0.5, 0.5))
for (i in (n1+1):n)
 freq[[i]] =  list(alleles = 1:2, afreq = c(MAF, 1-MAF))
x = setMarkers(x, locusAttributes = freq)


## Parent offspring
ids = c(1, 4); theta = 0; seed = 17
foo1 = compareCI(theta, x, ids, n, N, B, asymptotic = FALSE)

## half sibs
ids = c(4,5); theta = 0.5; seed = 17;
foo2 = compareCI(theta, x, ids, n, N, B, asymptotic= FALSE)

## Unrelated
ids = c(1,3); theta = 1
foo3 = compareCI(theta, x, ids, n, N, B, asymptotic = FALSE)

cbind(theta = rep(c(0,0.5,1), each = 2),
      rbind(foo1$average, foo2$average, foo3$average))

}
